sample_app_django/requirements.txt
asgiref==3.5.2
Django==4.1.4
pip==22.3.1
setuptools==65.6.3
sqlparse==0.4.3
tzdata==2022.7
wheel==0.38.4
# django-cors-headers==3.7.0


djangorestframework~=3.14.0